[v6.1.1] 整理 `zhengcheng-parent` 项目的系统模块 ums 权限系统调试 需要补充权限设计文档

+++ zhengcheng-micro/ums/src/main/java/com/zhengcheng/ums/controller/admin/system/SysUserController.java
@@ -3,12 +3,14 @@ package com.zhengcheng.ums.controller.admin.system;
 
 import com.zhengcheng.common.domain.PageResult;
 import com.zhengcheng.common.domain.Result;
+import com.zhengcheng.satoken.utils.SaTokenUtil;
 import com.zhengcheng.ums.domain.entity.SysRoleEntity;
 import com.zhengcheng.ums.domain.entity.SysUserEntity;
-import com.zhengcheng.ums.service.SysPermissionService;
+import com.zhengcheng.ums.service.SysPasswordService;
 import com.zhengcheng.ums.service.SysRoleService;
 import com.zhengcheng.ums.service.SysUserService;
 
+import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
@@ -41,7 +43,7 @@ public class SysUserController {
     @Autowired
     private SysRoleService roleService;
     @Autowired
-    private SysPermissionService permissionService;
+    private SysPasswordService passwordService;
 
     /**
      * 分页
@@ -56,6 +58,7 @@ public class SysUserController {
     /**
      * 信息
      */
+    @SaCheckPermission("system:user:query")
     @GetMapping({"{userId}", "/"})
     public Result getInfo(@PathVariable(value = "userId", required = false) Long userId) {
         Result res = Result.ok();
@@ -73,6 +76,7 @@ public class SysUserController {
     /**
      * 新增用户
      */
+    @SaCheckPermission("system:user:add")
     @PostMapping
     public Result add(@Validated @RequestBody SysUserEntity sysUserEntity) {
         userService.insertUser(sysUserEntity);
@@ -82,6 +86,7 @@ public class SysUserController {
     /**
      * 修改
      */
+    @SaCheckPermission("system:user:edit")
     @PutMapping
     public Result update(@Validated @RequestBody SysUserEntity user) {
         userService.checkUserAllowed(user);
@@ -100,17 +105,19 @@ public class SysUserController {
     /**
      * 删除
      */
+    @SaCheckPermission("system:user:remove")
     @DeleteMapping("/{userIds}")
     public Result remove(@PathVariable Long[] userIds) {
-//        if (ArrayUtils.contains(userIds, SecurityUtils.getUserId())) {
-//            return Result.error("当前用户不能删除");
-//        }
+        if (ArrayUtils.contains(userIds, SaTokenUtil.getUserId())) {
+            return Result.error("当前用户不能删除");
+        }
         return Result.ok(userService.deleteUserByIds(userIds));
     }
 
     /**
      * 根据用户编号获取授权角色
      */
+    @SaCheckPermission("system:user:query")
     @GetMapping("/authRole/{userId}")
     public Result authRole(@PathVariable("userId") Long userId) {
         Result res = Result.ok();
@@ -124,6 +131,7 @@ public class SysUserController {
     /**
      * 用户授权角色
      */
+    @SaCheckPermission("system:user:edit")
     @PutMapping("/authRole")
     public Result insertAuthRole(Long userId, Long[] roleIds) {
         if (!SysUserEntity.isAdmin(userId)) {
@@ -138,16 +146,18 @@ public class SysUserController {
     /**
      * 重置密码
      */
+    @SaCheckPermission("system:user:resetPwd")
     @PutMapping("/resetPwd")
     public Result resetPwd(@RequestBody SysUserEntity user) {
         userService.checkUserAllowed(user);
-//        user.setPassword(SecurityUtils.encryptPassword(user.getPassword()));
+        user.setPassword(passwordService.encode(user.getPassword()));
         return Result.ok(userService.resetPwd(user));
     }
 
     /**
      * 状态修改
      */
+    @SaCheckPermission("system:user:edit")
     @PutMapping("/changeStatus")
     public Result changeStatus(@RequestBody SysUserEntity user) {
         userService.checkUserAllowed(user);
